By Yi Zhang @June 14, 2012 During RHEL6.3 release
this checkin will still contain the following failures:
[ FAIL ] ipapassword_globalpolicy_classes_upperbound  <<< bug reported
[ FAIL ] ipapassword_grouppolicy_minlifetime_default
[ FAIL ] ipapassword_grouppolicy_history_default      <<< bug reported
[ FAIL ] ipapassword_nestedgrouppw_maxlife_conflict 

for minlife and maxlife testing, there is a weird error. Krb5kdc sometimes (randomly) report "CLIENT NOT FOUND" error, and there are coresponding SERC error 3 for it. The test case passes sometimes. I have separate test script to reproduce this behave. 
I will continue working on this issue.

Also in this checkin
1. greatly reduced randomly generated data usage. 
2. add more log message for troubleshooting purpose
3. recuded the logic complexity for test cases around classes





git-svn-id: https://svn.devel.redhat.com/repos/ipa-tests/trunk@3656 aeb79ed8-e03d-0410-be94-af332b95afa2